refactor(custom-templates): move variables field to react
