--- app/portainer/custom-templates/CustomTemplatesVariablesDefinitionField/index.ts
@@ -1,4 +1,0 @@
-export {
-  CustomTemplatesVariablesDefinitionField,
-  CustomTemplatesVariablesDefinitionFieldAngular,
-} from './CustomTemplatesVariablesDefinitionField';
