+++ app/portainer/views/custom-templates/create-custom-template-view/createCustomTemplateViewController.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { AccessControlFormData } from 'Portainer/components/accessControlForm/porAccessControlFormModel';
 import { TEMPLATE_NAME_VALIDATION_REGEX } from '@/constants';
-import { getTemplateVariables, intersectVariables } from '@/portainer/custom-templates/utils';
+import { getTemplateVariables, intersectVariables } from '@/react/portainer/custom-templates/components/utils';
 
 class CreateCustomTemplateViewController {
   /* @ngInject */
