+++ app/react/portainer/custom-templates/components/CustomTemplatesVariablesDefinitionField/index.ts
@@ -1,0 +1,1 @@
+export { CustomTemplatesVariablesDefinitionField } from './CustomTemplatesVariablesDefinitionField';
