+++ app/react/portainer/custom-templates/components/CustomTemplatesVariablesField/index.ts
@@ -1,0 +1,1 @@
+export { CustomTemplatesVariablesField } from './CustomTemplatesVariablesField';
