+++ app/react/portainer/custom-templates/components/utils.ts
